+++ boxsdk/object/events.py
@@ -93,9 +93,7 @@ class Events(BaseEndpoint):
         }
         box_response = self._session.get(url, params=params)
         response = box_response.json().copy()
-        if 'entries' in response:
-            response['entries'] = [self.translator.translate(item['type'])(item) for item in response['entries']]
-        return response
+        return self.translator.translate(self._session, response_object=response)
 
     @api_call
     def get_latest_stream_position(self, stream_type=UserEventsStreamType.ALL):