+++ boxsdk/object/file.py
@@ -1,6 +1,6 @@
 # coding: utf-8
 
-from __future__ import unicode_literals
+from __future__ import unicode_literals, absolute_import
 
 import json
 
@@ -15,6 +15,7 @@ class File(Item):
 
     _item_type = 'file'
 
+    @api_call
     def preflight_check(self, size, name=None):
         """
         Make an API call to check if the file can be updated with the new name and size of the file.
@@ -34,6 +35,36 @@ class File(Item):
             size=size,
             name=name,
             file_id=self._object_id,
+        )
+
+    def create_upload_session(self, file_size, file_name=None):
+        """
+        Create a new chunked upload session for uploading a new version of the file.
+
+        :param file_size:
+            The size of the file in bytes that will be uploaded.
+        :type file_size:
+            `int`
+        :param file_name:
+            The new name of the file version that will be uploaded.
+        :type file_name:
+            `unicode` or None
+        :returns:
+            A :class:`UploadSession` object.
+        :rtype:
+            :class:`UploadSession`
+        """
+        body_params = {
+            'file_id': self.object_id,
+            'file_size': file_size,
+        }
+        if file_name is not None:
+            body_params['file_name'] = file_name
+        url = self.get_url('upload_sessions').replace(self.session.api_config.BASE_API_URL, self.session.api_config.UPLOAD_URL)
+        response = self._session.post(url, data=json.dumps(body_params)).json()
+        return self.translator.translate(
+            session=self._session,
+            response_object=response,
         )
 
     def _get_accelerator_upload_url_for_update(self):
@@ -258,9 +289,11 @@ class File(Item):
             `unicode` or None
         :param unshared_at:
             The date on which this link should be disabled. May only be set if the current user is not a free user
-            and has permission to set expiration dates.
+            and has permission to set expiration dates.  Takes an RFC3339-formatted string, e.g.
+            '2018-10-31T23:59:59-07:00' for 11:59:59 PM on October 31, 2018 in the America/Los_Angeles timezone.
+            The time portion can be omitted, which defaults to midnight (00:00:00) on that date.
         :type unshared_at:
-            :class:`datetime.date` or None
+            `unicode` or None
         :param allow_preview:
             Whether or not the item being shared can be previewed when accessed via the shared link.
             If this parameter is None, the default setting will be used.
@@ -328,7 +361,7 @@ class File(Item):
             `unicode`
         """
         url = self._session.get_url('comments')
-        comment_class = self._session.translator.translate('comment')
+        comment_class = self._session.translator.get('comment')
         data = comment_class.construct_params_from_message(message)
         data['item'] = {
             'type': 'file',
@@ -336,12 +369,12 @@ class File(Item):
         }
         box_response = self._session.post(url, data=json.dumps(data))
         response = box_response.json()
-        return self._session.translator.translate(response['type'])(
+        return self._session.translator.translate(
             session=self._session,
-            object_id=response['id'],
             response_object=response,
         )
 
+    @api_call
     def create_task(self, message=None, due_at=None):
         """
         Create a task on the given file.
@@ -373,12 +406,12 @@ class File(Item):
             task_attributes['due_at'] = due_at
         box_response = self._session.post(url, data=json.dumps(task_attributes))
         response = box_response.json()
-        return self.translator.translate(response['type'])(
+        return self.translator.translate(
             session=self._session,
-            object_id=response['id'],
             response_object=response,
         )
 
+    @api_call
     def get_tasks(self, fields=None):
         """
         Get the entries in the file tasks.