+++ boxsdk/object/folder.py
@@ -91,6 +91,7 @@ class Folder(Item):
 
     _item_type = 'folder'
 
+    @api_call
     def preflight_check(self, size, name):
         """
         Make an API call to check if a new file with given name and size can be uploaded to this folder.
@@ -110,6 +111,35 @@ class Folder(Item):
             size=size,
             name=name,
             parent_id=self._object_id,
+        )
+
+    def create_upload_session(self, file_size, file_name):
+        """
+        Creates a new chunked upload session for upload a new file.
+
+        :param file_size:
+            The size of the file in bytes that will be uploaded.
+        :type file_size:
+            `int`
+        :param file_name:
+            The name of the file that will be uploaded.
+        :type file_name:
+            `unicode`
+        :returns:
+            A :class:`UploadSession` object.
+        :rtype:
+            :class:`UploadSession`
+        """
+        url = '{0}/files/upload_sessions'.format(self.session.api_config.UPLOAD_URL)
+        body_params = {
+            'folder_id': self.object_id,
+            'file_size': file_size,
+            'file_name': file_name,
+        }
+        response = self._session.post(url, data=json.dumps(body_params)).json()
+        return self.translator.translate(
+            session=self._session,
+            response_object=response,
         )
 
     def _get_accelerator_upload_url_fow_new_uploads(self):
@@ -255,10 +285,8 @@ class Folder(Item):
         file_response = self._session.post(url, data=data, files=files, expect_json_response=False).json()
         if 'entries' in file_response:
             file_response = file_response['entries'][0]
-        file_id = file_response['id']
-        return self.translator.translate(file_response['type'])(
+        return self.translator.translate(
             session=self._session,
-            object_id=file_id,
             response_object=file_response,
         )
 
@@ -337,9 +365,8 @@ class Folder(Item):
         }
         box_response = self._session.post(url, data=json.dumps(data))
         response = box_response.json()
-        return self.__class__(
+        return self.translator.translate(
             session=self._session,
-            object_id=response['id'],
             response_object=response,
         )
 
@@ -409,13 +436,12 @@ class Folder(Item):
         params = {'notify': notify}
         box_response = self._session.post(url, expect_json_response=True, data=data, params=params)
         collaboration_response = box_response.json()
-        collab_id = collaboration_response['id']
-        return self.translator.translate(collaboration_response['type'])(
+        return self.translator.translate(
             session=self._session,
-            object_id=collab_id,
             response_object=collaboration_response,
         )
 
+    @api_call
     def create_web_link(self, target_url, name=None, description=None):
         """
         Create a WebLink with a given url.
@@ -449,9 +475,8 @@ class Folder(Item):
         if description is not None:
             web_link_attributes['description'] = description
         response = self._session.post(url, data=json.dumps(web_link_attributes)).json()
-        return self.translator.translate(response['type'])(
+        return self.translator.translate(
             session=self._session,
-            object_id=response['id'],
             response_object=response
         )