+++ boxsdk/object/storage_policy_assignment.py
@@ -1,0 +1,9 @@
+# coding:utf-8
+from __future__ import unicode_literals, absolute_import
+from .base_object import BaseObject
+
+
+class StoragePolicyAssignment(BaseObject):
+    """Represents the storage policy assignment"""
+
+    _item_type = 'storage_policy_assignment'